Refactor stepping_mode.rs from debug/mod.rs
